Takudo (#53)
 coupon-code, filter shop, update shop item, update amount

+++ database/migrations/2022_04_27_170457_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('price');
-            $table->float('amount');
+            $table->integer('amount');
             $table->string('image')->nullable();
             $table->text('description');
             $table->bigInteger('category_id');
